Global text size fix

+++ app/src/main/java/com/lardis/ivan/testcustomview/graphview/helpers/HelperLayoutClass.java
@@ -76,7 +76,7 @@ public class HelperLayoutClass {
         if (minTextSize > maxAllowedTextSize)
             paint.setTextSize(maxAllowedTextSize);
         else
-            paint.setTextSize(minTextSize);
+            paint.setTextSize((int) minTextSize);
     }
 
     public static int getScreenWidth(Context context)
